:bug: fix build issues on frontend
