+++ src/actions/completeActivityAction.ts
@@ -2,7 +2,7 @@
 
 import "server-only";
 
-import { auth } from "@/actions/authActions";
+import { auth, signOut } from "@/actions/authActions";
 import { sendInvalidationCodes } from "@/revalidation/sendInvalidationCode";
 import { completeTeamActivity } from "@/services/completeActivityService";
 import { getActivityById } from "@/services/getActivityByIdService";
@@ -18,10 +18,6 @@ export async function completeActivityAction(
   activityCode: string,
 ): Promise<void> {
   const { teamId } = await auth();
-
-  if (!teamId) {
-    throw new Error("Unauthenticated");
-  }
 
   if (!activityCode || activityCode.length !== 6) {
     throw new Error("Incorrect activity code");
