--- src/app/(auth)/layout.tsx
@@ -1,11 +1,0 @@
-import { protect } from "@/actions/authActions";
-
-export default async function AuthLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  await protect();
-
-  return <>{children}</>;
-}
