+++ src/app/board/_components/ActivityDrawer.tsx
