+++ src/app/board/_components/BingoBoardSkeleton.tsx
